+++ lib/openapi_first/router_required.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module OpenapiFirst
+  module RouterRequired
+    def call(env)
+      unless env.key?(OPERATION)
+        raise 'OpenapiFirst::Router missing in middleware stack. Did you forget adding OpenapiFirst::Router?'
+      end
+
+      super
+    end
+  end
+end